requirements.txt
fastapi
uvicorn[standard]
anthropic
vonage
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
supabase
httpx
python-dotenv
